separate abstract base operand class from its implementations
