+++ parser_library/src/context/ordinary_assembly/postponed_statement.cpp
@@ -20,7 +20,7 @@ namespace hlasm_plugin::parser_library::context {
 
 const processing::postponed_statement_impl* postponed_statement::impl() const
 {
-    return dynamic_cast<const processing::postponed_statement_impl*>(this);
+    return static_cast<const processing::postponed_statement_impl*>(this);
 }
 
 } // namespace hlasm_plugin::parser_library::context
