+++ parser_library/src/context/ordinary_assembly/postponed_statement.h
@@ -38,4 +38,5 @@ using post_stmt_ptr = std::unique_ptr<postponed_statement>;
 
 } // namespace hlasm_plugin::parser_library::context
 
+
 #endif
