+++ parser_library/test/semantics/operand_test.cpp
@@ -14,7 +14,7 @@
 
 #include "gmock/gmock.h"
 
-#include "semantics/operand.h"
+#include "semantics/operand_impls.h"
 
 using namespace hlasm_plugin::parser_library::semantics;
 using namespace hlasm_plugin::parser_library;
